requirements.txt
torchvision == 0.16.0
tqdm
pandas